Introduce shared settings list template
