+++ app/src/main/java/com/glancemap/glancemapwearos/presentation/features/settings/SettingsScreen.kt
@@ -1,18 +1,13 @@
 package com.glancemap.glancemapwearos.presentation.features.settings
 
 import android.widget.Toast
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Gavel
 import androidx.compose.material.icons.filled.UnfoldMore
 import androidx.compose.runtime.*
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.navigation.NavHostController
-import androidx.wear.compose.foundation.lazy.ScalingLazyColumn
 import androidx.wear.compose.material3.AlertDialog
 import androidx.wear.compose.material3.Button
 import androidx.wear.compose.material3.MaterialTheme
@@ -22,7 +17,6 @@ import com.glancemap.glancemapwearos.presentation.features.gpx.GpxViewModel
 import com.glancemap.glancemapwearos.presentation.features.maps.MapViewModel
 import com.glancemap.glancemapwearos.presentation.navigation.WatchRoutes
 import com.google.android.horologist.annotations.ExperimentalHorologistApi
-import com.google.android.horologist.compose.layout.ScreenScaffold
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -48,7 +42,6 @@ fun SettingsScreen(
     var partialSummary by remember {
         mutableStateOf(StalePartialTransferCleaner.PartialFilesSummary(count = 0, totalBytes = 0L))
     }
-    val listState = rememberSettingsScalingLazyListState(topPadding = listTokens.topPadding)
     val isMetric by viewModel.isMetric.collectAsState()
     val backButtonExitsNavigation by viewModel.backButtonExitsNavigation.collectAsState()
     val unitOptions =
@@ -82,140 +75,125 @@ fun SettingsScreen(
             else -> "${partialSummary.count} files · ${formatStorageSize(partialSummary.totalBytes)}"
         }
 
-    ScreenScaffold(scrollState = listState) {
-        ScalingLazyColumn(
-            modifier = Modifier.fillMaxSize(),
-            state = listState,
-            contentPadding =
-                PaddingValues(
-                    start = listTokens.horizontalPadding,
-                    end = listTokens.horizontalPadding,
-                    top = listTokens.topPadding,
-                    bottom = listTokens.bottomPadding,
-                ),
-            verticalArrangement = Arrangement.spacedBy(listTokens.itemSpacing),
-            anchorType = SettingsListAnchorType,
-            autoCentering = SettingsListAutoCentering,
-        ) {
-            item {
-                Text(
-                    "General",
-                    style = MaterialTheme.typography.titleMedium,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-            }
-            item {
-                SettingsPickerChip(
-                    label = "Units",
-                    secondaryLabel = if (isMetric) "Metric" else "Imperial",
-                    iconImageVector = Icons.Filled.UnfoldMore,
-                    onClick = { showUnitsPicker = true },
-                )
-            }
-            item {
-                SettingsSectionChip(
-                    label = "GPS settings",
-                    onClick = { navController.navigate(WatchRoutes.GPS_SETTINGS) },
-                )
-            }
-            item {
-                SettingsSectionChip(
-                    label = "Compass settings",
-                    onClick = { navController.navigate(WatchRoutes.COMPASS_SETTINGS) },
-                )
-            }
+    WearSettingsListScreen(listTokens = listTokens) {
+        item {
+            Text(
+                "General",
+                style = MaterialTheme.typography.titleMedium,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+            )
+        }
+        item {
+            SettingsPickerChip(
+                label = "Units",
+                secondaryLabel = if (isMetric) "Metric" else "Imperial",
+                iconImageVector = Icons.Filled.UnfoldMore,
+                onClick = { showUnitsPicker = true },
+            )
+        }
+        item {
+            SettingsSectionChip(
+                label = "GPS settings",
+                onClick = { navController.navigate(WatchRoutes.GPS_SETTINGS) },
+            )
+        }
+        item {
+            SettingsSectionChip(
+                label = "Compass settings",
+                onClick = { navController.navigate(WatchRoutes.COMPASS_SETTINGS) },
+            )
+        }
 
-            item { Text("Screen settings", style = MaterialTheme.typography.titleMedium) }
-            item {
-                SettingsSectionChip(
-                    label = "POI settings",
-                    onClick = { navController.navigate(WatchRoutes.POI_SETTINGS) },
-                )
-            }
-            item {
-                SettingsSectionChip(
-                    label = "GPX settings",
-                    onClick = { navController.navigate(WatchRoutes.GPX_SETTINGS) },
-                )
-            }
-            item {
-                SettingsSectionChip(
-                    label = "Map settings",
-                    onClick = { navController.navigate(WatchRoutes.MAP_SETTINGS) },
-                )
-            }
-            item {
-                DownloadSettingsSectionChip(
-                    onClick = { navController.navigate(WatchRoutes.DOWNLOAD_SETTINGS) },
-                )
-            }
+        item { Text("Screen settings", style = MaterialTheme.typography.titleMedium) }
+        item {
+            SettingsSectionChip(
+                label = "POI settings",
+                onClick = { navController.navigate(WatchRoutes.POI_SETTINGS) },
+            )
+        }
+        item {
+            SettingsSectionChip(
+                label = "GPX settings",
+                onClick = { navController.navigate(WatchRoutes.GPX_SETTINGS) },
+            )
+        }
+        item {
+            SettingsSectionChip(
+                label = "Map settings",
+                onClick = { navController.navigate(WatchRoutes.MAP_SETTINGS) },
+            )
+        }
+        item {
+            DownloadSettingsSectionChip(
+                onClick = { navController.navigate(WatchRoutes.DOWNLOAD_SETTINGS) },
+            )
+        }
 
-            item { Text("Advanced settings", style = MaterialTheme.typography.titleMedium) }
-            item {
-                SettingsSectionChip(
-                    label = "Debugging",
-                    onClick = { navController.navigate(WatchRoutes.DEBUG_SETTINGS) },
-                )
-            }
-            item {
-                SettingsToggleChip(
-                    checked = backButtonExitsNavigation,
-                    onCheckedChanged = viewModel::setBackButtonExitsNavigation,
-                    label = "Exit back button",
-                    secondaryLabel = "For compatible watches",
-                )
-            }
-            item {
-                SettingsPickerChip(
-                    label = if (isClearingCache) "Clearing cache..." else "Clear cache",
-                    iconImageVector = null,
-                    onClick = {
-                        if (!isClearingCache) {
-                            showClearCacheDialog = true
-                        }
+        item { Text("Advanced settings", style = MaterialTheme.typography.titleMedium) }
+        item {
+            SettingsSectionChip(
+                label = "Debugging",
+                onClick = { navController.navigate(WatchRoutes.DEBUG_SETTINGS) },
+            )
+        }
+        item {
+            SettingsToggleChip(
+                checked = backButtonExitsNavigation,
+                onCheckedChanged = viewModel::setBackButtonExitsNavigation,
+                label = "Exit back button",
+                secondaryLabel = "For compatible watches",
+            )
+        }
+        item {
+            SettingsPickerChip(
+                label = if (isClearingCache) "Clearing cache..." else "Clear cache",
+                iconImageVector = null,
+                onClick = {
+                    if (!isClearingCache) {
+                        showClearCacheDialog = true
+                    }
+                },
+            )
+        }
+        item {
+            SettingsPickerChip(
+                label =
+                    if (isClearingPartialFiles) {
+                        "Clearing partial transfer..."
+                    } else {
+                        "Clear partial transfer"
                     },
-                )
-            }
-            item {
-                SettingsPickerChip(
-                    label =
-                        if (isClearingPartialFiles) {
-                            "Clearing partial transfer..."
-                        } else {
-                            "Clear partial transfer"
-                        },
-                    secondaryLabel = partialSummaryText,
-                    iconImageVector = null,
-                    onClick = {
-                        if (isClearingPartialFiles) return@SettingsPickerChip
-                        if (partialSummary.count <= 0) {
-                            Toast
-                                .makeText(
-                                    context,
-                                    "No partial transfer files",
-                                    Toast.LENGTH_SHORT,
-                                ).show()
-                        } else {
-                            showClearPartialTransfersDialog = true
-                        }
-                    },
-                )
-            }
-            item {
-                SettingsPickerChip(
-                    label = "Reset to Default",
-                    iconImageVector = null,
-                    onClick = { navController.navigate(WatchRoutes.RESET_DEFAULTS_CONFIRM) },
-                )
-            }
-            item {
-                SettingsSectionChip(
-                    label = "Credits & Legal",
-                    iconImageVector = Icons.Filled.Gavel,
-                    onClick = { navController.navigate(WatchRoutes.LICENSES) },
-                )
-            }
+                secondaryLabel = partialSummaryText,
+                iconImageVector = null,
+                onClick = {
+                    if (isClearingPartialFiles) return@SettingsPickerChip
+                    if (partialSummary.count <= 0) {
+                        Toast
+                            .makeText(
+                                context,
+                                "No partial transfer files",
+                                Toast.LENGTH_SHORT,
+                            ).show()
+                    } else {
+                        showClearPartialTransfersDialog = true
+                    }
+                },
+            )
+        }
+        item {
+            SettingsPickerChip(
+                label = "Reset to Default",
+                iconImageVector = null,
+                onClick = { navController.navigate(WatchRoutes.RESET_DEFAULTS_CONFIRM) },
+            )
+        }
+        item {
+            SettingsSectionChip(
+                label = "Credits & Legal",
+                iconImageVector = Icons.Filled.Gavel,
+                onClick = { navController.navigate(WatchRoutes.LICENSES) },
+            )
         }
     }
